# api.yml
openapi: 3.0.0
info:
  title: HK News API
  description: Serves aggregated news from 13 local news publishers in Hong Kong
  version: 1.1.0
  contact:
    name: GitHub
    url: https://github.com/ayltai/hknews
  license:
    name: MIT
    url: https://github.com/ayltai/hknews/blob/master/LICENSE
servers:
  - url: https://hknews.dev
paths:
  /sources:
    get:
      summary: Gets a list of news publishers
      description: Gets a list of news publishers supported by the server
      responses:
        200:
          description: The requested list of news publishers was retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
      tags:
        - news publishers
  /items/{sourceNames}/{categoryNames}/{days}:
    get:
      summary: Gets a list of news records
      description: Gets a list of news records of specific news categories and from specific news publishers
      parameters:
        - in: path
          name: sourceNames
          description: A list of news publishers, separated by commas
          schema:
            type: string
            example: 蘋果日報,香港電台
          required: true
        - in: path
          name: categoryNames
          description: A list of news categories, separated by commas
          schema:
            type: string
            example: 港聞,國際
          required: true
        - in: path
          name: days
          description: The number of previous days of news to retrieve
          schema:
            type: string
            example: 2
          required: true
        - in: query
          name: page
          description: Page offset
          schema:
            type: integer
            default: 0
          required: false
        - in: query
          name: size
          description: Page size
          schema:
            type: integer
            default: 100
          required: false
      responses:
        200:
          description: The requested list of news records was retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageOfItems'
        400:
          description: At least one of the request parameters was invalid
      tags:
        - news
  /item/{id}:
    get:
      summary: Gets a news record
      description: Gets a specific news record by its ID
      parameters:
        - in: path
          name: id
          description: An unique ID that represents the news record to be retrieved
          schema:
            type: string
            format: uuid
            example: 3c765a4f-4cf9-46e7-8331-6ed6208c9644
          required: true
      responses:
        200:
          description: The requested news record was retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
      tags:
        - news
components:
  schemas:
    Sort:
      type: object
      required:
        - sorted
        - unsorted
        - empty
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
    Pageable:
      type: object
      required:
        - sort
        - offset
        - pageSize
        - pageNumber
        - unpaged
        - paged
      properties:
        sort:
          $ref: '#/components/schemas/Sort'
        offset:
          type: integer
          description: The offset to be taken according to the underlying page and page size
        pageSize:
          type: integer
          description: The number of items to be returned
        pageNumber:
          type: integer
          description: The page to be returned
        unpaged:
          type: boolean
          description: Whether the current object does not contain pagination information
        paged:
          type: boolean
          description: Whether the current object contains pagination information
    PageOfItems:
      allOf:
        - type: object
          required:
            - content
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Item'
        - $ref: '#/components/schemas/Page'
    Item:
      type: object
      required:
        - id
        - url
        - publishDate
        - sourceName
        - categoryName
        - images
        - videos
      properties:
        id:
          type: string
          format: uuid
          description: The unique ID that represents a news record
          example: 3c765a4f-4cf9-46e7-8331-6ed6208c9644
        title:
          type: string
          description: The news title
        description:
          type: string
          description: The news detailed description
        url:
          type: string
          description: The URL where the news was retrieved from
        publishDate:
          type: string
          format: date-time
          description: The date and time when the news record was first published
        sourceName:
          type: string
          description: The name of the news publisher who published the news record
          example: 蘋果日報
        categoryName:
          type: string
          description: The news category that the news record belongs to
          example: 港聞
        images:
          type: array
          items:
            $ref: '#/components/schemas/Image'
        videos:
          type: array
          items:
            $ref: '#/components/schemas/Video'
    Image:
      type: object
      required:
        - imageUrl
      properties:
        imageUrl:
          type: string
          description: The URL of the image that belongs to a news record
        description:
          type: string
          description: The text description associated with the image
    Video:
      type: object
      required:
        - videoUrl
        - imageUrl
      properties:
        videoUrl:
          type: string
          description: The URL of the video that belongs to a news record
        imageUrl:
          type: string
          description: The URL of the video thumbnail that belongs to a news record
    Page:
      type: object
      required:
        - pageable
        - sort
        - number
        - size
        - totalPages
        - totalElements
        - numberOfElements
        - first
        - last
        - empty
      properties:
        pageable:
          $ref: '#/components/schemas/Pageable'
        sort:
          $ref: '#/components/schemas/Sort'
        number:
          type: integer
          description: The number of current page
        size:
          type: integer
          description: The size of the page
        totalPages:
          type: integer
          description: The number of total pages
        totalElements:
          type: integer
          description: The total amount of elements
        numberOfElements:
          type: integer
          description: The number of elements currently on this page
        first:
          type: boolean
          description: Whether the current page is the first one
        last:
          type: boolean
          description: Whether the current page is the last one
        empty:
          type: boolean
          description: Whether the current page is empty
    Category:
      type: object
      required:
        - name
        - urls
      properties:
        name:
          type: string
          description: The name of the news category
          example: 港聞
        urls:
          type: array
          items:
            type: string
            description: The URL where news of specific category can be retrieved from
            example: https://hk.appledaily.com/pf/api/v3/content/fetch/query-feed?query={"feedOffset":0,"feedQuery":"taxonomy.primary_section._id:\"/daily/local\"+AND+type:story+AND+publish_date:[now-48h/h+TO+now]","feedSize":100,"sort":"display_date:desc"}
    Source:
      type: object
      required:
        - name
        - categories
      properties:
        name:
          type: string
          description: The name of the news publisher
          example: 蘋果日報
        imageUrl:
          type: string
          description: A relative URL pointing to the logo image of the news publisher
          example: /images/appledaily.png
        categories:
          type: array
          description: A list of news categories supported by the news publisher
          items:
            $ref: '#/components/schemas/Category'
tags:
  - name: news
    description: News records related operations
  - name: news publishers
    description: News publishers related operations